requirements.txt
numpy>=1.23
pandas>=2.0
scikit-learn>=1.3
imbalanced-learn>=0.11
xgboost>=1.7
lightgbm>=4.0
catboost>=1.2
shap>=0.44
optuna>=3.4
mlflow>=2.10
matplotlib>=3.7
seaborn>=0.12
plotly>=5.18
streamlit>=1.32
scipy>=1.10
pyarrow>=14.0

# Encoding & config
category-encoders>=2.6
pyyaml>=6.0

# Optional interpretability
lime>=0.2.0
